refactor(weave): merge session _redaction helpers into pii_redaction
Per review feedback (PR #6846): consolidate the Session SDK redaction
wrappers into weave/utils/pii_redaction.py alongside the primitives they
route through. Delete weave/session/_redaction.py; session.py and its
tests now import from pii_redaction.

- Move redact_string, _restore_literals, redact_messages,
  redact_system_instructions from _redaction.py into pii_redaction.py.
- Switch the optional-dep guard from per-function lazy imports to a
  module-level try/except + _HAS_PRESIDIO flag; functions check the flag
  and raise the install hint on miss.
- Update test mocks from weave.session._redaction.redact_pii to
  weave.utils.pii_redaction.redact_pii.

233 session tests pass. Ruff check + format clean.
